Impl을 삭제하고 service로만 구현
